Created new component CustomerForm that will share the logic for edit and new form pages
